backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
SQLAlchemy==2.0.34
psycopg[binary]==3.2.9
python-dotenv==1.0.1
Werkzeug==3.0.3

